requirements.txt
discord.py>=2.4.0
python-dotenv>=1.0.0
yt-dlp>=2024.12.13
PyNaCl>=1.5.0
aiohttp>=3.10.0
